Added a second disk and set boot order.
We need a second disk in order to install CoreOS to.

In addition to that, we should make sure to first boot into that "fresh"
disk first when possible to make sure that after reboot we boot into the
newly written disk.

Cluster was installed successfully and agent isn't present anymore but I
don't have my changes on the node becuase I forgot to apply the
machine-config object to the cluster.

Need to add a MachineConfig manifest to the installation with the
`osImageURL` set to the customized container image.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>